Extracted fetchRandomMediaContent to separate file
